Renamed config.json to keys.json and added check for env var if keys.json isn't found
