+++ .gitignore
@@ -4,4 +4,4 @@ package-lock.json
 scratch.js
 bots/**/action-code/**
 bots/**/
-config.json
+keys.json
