+++ README.md
@@ -14,7 +14,7 @@ This project allows an AI model to write/execute code on your computer that may 
 
 ## Installation
 
-Rename `config.example.json` to `config.json` and fill in the desired API keys
+Rename `keys.example.json` to `keys.json` and fill in the desired API keys
 | API | Config Variable | Example Model name | Docs |
 |------|------|------|------|
 | OpenAI | `OPENAI_API_KEY` | `gpt-3.5-turbo` | [docs](https://platform.openai.com/docs/models) | (optionally add `OPENAI_ORG_ID`)
