+++ src/models/claude.js
@@ -1,5 +1,5 @@
 import Anthropic from '@anthropic-ai/sdk';
-import configJson from "../../config.json" assert { type: "json" };
+import configJson from "../../keys.json" assert { type: "json" };
 
 
 export class Claude {
@@ -11,8 +11,10 @@ export class Claude {
             config.baseURL = url;
         if (configJson.ANTHROPIC_API_KEY)
             config.apiKey = configJson.ANTHROPIC_API_KEY;
+        else if (process.env.ANTHROPIC_API_KEY)
+            config.apiKey = process.env.ANTHROPIC_API_KEY;
         else
-            throw new Error('Anthropic API key missing! Make sure you set your ANTHROPIC_API_KEY in your config.json.');
+            throw new Error('Anthropic API key missing! Make sure you set your ANTHROPIC_API_KEY in your keys.json.');
 
         this.anthropic = new Anthropic(config);
     }
