+++ src/models/gemini.js
@@ -1,16 +1,18 @@
 import { GoogleGenerativeAI } from '@google/generative-ai';
 import { toSinglePrompt } from '../utils/text.js';
-import configJson from "../../config.json" assert { type: "json" };
+import configJson from "../../keys.json" assert { type: "json" };
 
 export class Gemini {
     constructor(model_name, url) {
         this.model_name = model_name;
         this.url = url;
 
-        if (!configJson.GEMINI_API_KEY) {
-            throw new Error('Gemini API key missing! Make sure you set your GEMINI_API_KEY in your config.json.');
-        }
-        this.genAI = new GoogleGenerativeAI(configJson.GEMINI_API_KEY);
+        if (configJson.GEMINI_API_KEY)
+            this.genAI = new GoogleGenerativeAI(configJson.GEMINI_API_KEY);
+        else if (process.env.GEMINI_API_KEY)
+            this.genAI = new GoogleGenerativeAI(process.env.GEMINI_API_KEY);
+        else
+            throw new Error('Gemini API key missing! Make sure you set your GEMINI_API_KEY in your keys.json.');
     }
 
     async sendRequest(turns, systemMessage) {
