+++ src/models/replicate.js
@@ -1,6 +1,6 @@
 import Replicate from 'replicate';
 import { toSinglePrompt } from '../utils/text.js';
-import configJson from "../../config.json" assert { type: "json" };
+import configJson from "../../keys.json" assert { type: "json" };
 
 // llama, mistral
 export class ReplicateAPI {
@@ -12,13 +12,16 @@ export class ReplicateAPI {
 			console.warn('Replicate API does not support custom URLs. Ignoring provided URL.');
 		}
 
-		if (!configJson.REPLICATE_API_KEY) {
-			throw new Error('Replicate API key missing! Make sure you set your REPLICATE_API_KEY in your config.json.');
-		}
-
-		this.replicate = new Replicate({
-			auth: configJson.REPLICATE_API_KEY,
-		});
+		if (configJson.REPLICATE_API_KEY)
+			this.replicate = new Replicate({
+				auth: configJson.REPLICATE_API_KEY,
+			});
+        else if (process.env.REPLICATE_API_KEY)
+			this.replicate = new Replicate({
+				auth: process.env.REPLICATE_API_KEY,
+			});
+        else
+            throw new Error('Replicate API key missing! Make sure you set your REPLICATE_API_KEY in your keys.json.');
 	}
 
 	async sendRequest(turns, systemMessage) {
